crud autor completo

+++ app/src/main/java/com/example/controlinventario/Libro/LibroDao.java
@@ -13,7 +13,7 @@ public interface LibroDao extends GenericDAO<LibroEntity> {
     List<LibroEntity> getAll();
 
     @Query("SELECT * FROM Libro WHERE idLibro = :id")
-    LibroEntity findById(int id);
+    LibroEntity findById(Long id);
 
     @Query("Select l.IDMATERIA from Libro l  order by l.IDMATERIA desc limit 1")
     Long getLastId();
